web/web06/src/requirements.txt
uwsgi==2.0.21
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
bcrypt==4.0.1
flask-session2==1.3.1